requirements.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
matplotlib==3.9.2
scikit-learn==1.5.1
pyyaml==6.0.2
tqdm==4.66.5

torch==2.4.1
torchvision==0.19.1

jupyter==1.1.1
notebook==7.2.2
